tests: move checkpoint tests into separate module
